backend/requirements-staging.txt
Django==3.1.3
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-otp==1.0.2
django-request-logging==0.7.2
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
psycopg2-binary==2.8.6
Pillow==8.0.1
PyJWT==1.7.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
typed-ast==1.4.1
wrapt==1.12.1
asgiref==3.3.1
astroid==2.4.2
drf-url-filters==0.5.1
gunicorn==20.0.4
drf-writable-nested==0.6.2
drf-spectacular==0.12.0
whitenoise==5.2.0
django-storages==1.11.1
boto3==1.16.63
requests==2.25.1
django-filter==2.4.0
django-model-utils==4.1.1
django-redis==4.12.1

# celery
celery==5.0.5
redis==3.5.3
